Adding support for authentication via session (#2)
* verifies the return of Get and error out in case of failure

* verifies the return of Get and error out in case of failure

* We already have the ID coming from common.types

* We already have the ID coming from common.types

* Makes client return the ApiClient object. Since we don't have the token
at this point, this makes possible to issue a request to retrieve the
token and later update the ApiClient struct. After we can still pass it
as common.Client

* SessionService only host a link to Sessions, to deal with Session in
reality you need to use the /Sessions endpoint. I've updated the Session
and Service to reflect that.

* Adds CreateSession and DeleteSession.
I'm not sure wheter these two should be inside of service. So I'm trying
to follow the same schema from the old calls.

* Adds a generic delete method, it already works for sessions.

* update Client interface to support the Delete()

* Adds DeleteSession

* adds a generic Post to ApiClient

* Post now is required as part of the Client interface

* We can now create new sessions and use those as part of the library.
Since the response payload from redfish doesn't contain the token. I'm
using the headers to polulate the auth information needed.

* adds examples for session and chassis queries

* update README.md with new connection schema