Add client connection settings
This changes the way you connect to a redfish service. ConnectDefault
will use default values to create an unauthenticated connection to the
service (typically useful for testing against the simulator), otherwise
you create a ClientConfig variable that contains all of your connection
settings and pass that in to a Connect call.

Closes: #33