Update numeric fields to use float
The Redfish schema uses the types 'integer' and 'numeric' for field
types. The integer type should be a go int, but numeric types could
contain decimal precision. These fields either need to use float32 or
float64 to handle this.

This updates all redfish objects to use a float type for any that are
defined as numeric in the schema.

Closes: #68

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>