Add chassis test for redfish_exporter failure
A user of redfish_exporter is getting a failure with one of their
Supermicro servers, but not all. Looking at the raw json returned from
these machines, the failure happens on a host that has a RAID chassis in
addition to the normal system chassis.

The results from getting the RAID chassis don't have as many values as
the full chassis, but according to the required properties defined in
the spec, it does return everything that is required.

To test to see if this was an issue in gofish, added a unit test that
would use the response from this system's RAID chassis and a new test
that validates only the spec-required minimum values. It appears gofish
handles this appropriately, but adding this test just in case. We may
want to add minimum required value tests for more objects, but this is a
start.

This is to help debug:
https://github.com/jenningsloy318/redfish_exporter/issues/15

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>