Hide ComputerSystem.HostedServices property for now
This property is actually an object that contains properties that link
to other services. The existing definition as a string is not correct,
so even though we don't have full support for it yet, at least mark it
private so that marshalling does not have problems if a host returns the
correct data structure.

Closes: #109

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>