Capture Bios etag for later updates
This updates our handling of an instances etag so we capture it on
initial retrieval of the object. Then this value can be used later for
any update calls to enforce that the object we are trying to update has
not been modified since our local version of it was retrieved.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>