Make collection retrievals parallel
This makes calls to get the objects from a collection link parallel.
Some collections can contain a large number of items to retrieve. Doing
this serially can have a noticable performance impact. This change
allows three fetches in parallel so we don't overwhelm the Redfish
service and cause a denial of service. With three concurrent operations
we can be processing some responses while queuing up the more, making it
a lot faster to process.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>
Co-Authored-by: Lorenzo Bolognesi <lorenzo.bolognesi@dxc.com>