Add deprecation note to Storage.StorageControllers
This property has been deprecated in favor of linking to a collection of
objects, rather than embedding in the storage object. This adds a note
to make it clear to consumers that this property may not be the right
way to access this information if talking to a newer system.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>