Fix Chassis controls link definition
The Controls property of the Chassis object was incorrectly defined as a
`common.Links` type, but it is actually a link to a collection, so it
should be a `common.Link` (singular). This changes it to the Link type
to avoid a JSON marshalling error on systems that populate this
property.

This also uncovered that the link that is used later to retrieve these
controls was never actually being set. Both issues are now fixed.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>