ListReferencedX: return ordered results.
Due to the parallelization of the collection's links retrieval, the
returned list of the ListReferencedX funcs cannot be trusted to be
ordered. This commit ensures that the returned list keeps the same order
as the collection's members list.

Update common.GetCollection() to return the list of the collection's
links in the same order they were retrieved.

Then, ListReferencedX funcs  use the original link list to reorder the
results before returning them. This proposal uses a helper map to store
the unordered resultsDue to the parallelization of the collection's
links retrieval, the returned list of the ListReferencedX funcs cannot
be trusted to be ordered. This commit ensures that the returned list
keeps the same order as the collection's members list..

I've updated all affected packages plus the template in the tools folder.