Use common generic fetch of collection objects
This moves the minimum supported Go version to 1.18 to allow the use of
gneerics in the code.

This adds a common function for getting all objects in a collection
referenced by a link. This is used by most objects to retrieve a set of
all objects of the given type. With the ability to use generics now,
this eliminates a lot of repetitive code.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>