Add common concurrent function to get list of objects (#354)
We have concurrency when getting a collection that contains a set of
links to objects. But some object properties are a list themselves and
not a Collection object. This adds a common function to be used for
these types of properties to take advantage of the concurrent processing
to fetch these objects from the system.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>