Update SMC OEM readwrite field values to match SMC documentation.
This commit *adds* more SMC entity field names that should be allowed for read/write.
This started as matching the field names provided in
https://www.supermicro.com/manuals/other/redfish-ref-guide-html/Content/general-content/bmc-configuration-examples.htm
but then continued to follow the JSON field names that were found higher in each file
as that matched what a new SMC chassis expected.

The previous field names were left as read/write field names in case some of the older chassis BMCs
expect those names.

The NTP and Syslog file changes have been confirmed to fix bugs that we've seen
on our end. All of the changes are additions rather than updates/renames, so this should
break nothing.