Initialize TLSClientConfig when using default transport
To allow insecure (unverified TLS) connections, we need to set a flag
allowing it. When using the default http client, or if a supplied client
has not been fully configured, the struct containing this flag is not
initialized and set to nil in the http client transport.

This makes sure the config property is initialized if we are configured
to set the flag to avoid nil pointer errors.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>